requirements.txt
pyrogram==2.0.106
TgCrypto==1.2.5
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
aiohttp==3.9.5
